+++ next-app/src/components/common/Layout/Layout.tsx
@@ -1,4 +1,4 @@
-import Nav from '../Nav'
+import Nav from './Nav'
 import { Container } from './Layout.style'
 import { useMemo } from 'react'
 import { useRouter } from 'next/router'
