+++ next-app/src/components/common/Layout/Nav/Nav.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 
-import TopNavBar from '../TopNavBar'
-import SideMenuBar from '../SideNavBar'
+import TopNavBar from './TopNavBar'
+import SideMenuBar from './SideNavBar'
 
 const Nav = () => {
   const [showSideBar, setShowSideBar] = useState<boolean | null>(null)
